backend/requirements.txt
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.1
asttokens==3.0.0
billiard==4.2.4
brotli==1.2.0
CacheControl==0.14.4
celery==5.5.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==46.0.5
cssselect2==0.9.0
Django==5.1.6
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-timezone-field==7.2.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
firebase-admin==6.9.0
fonttools==4.61.1
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.78.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
kombu==5.5.4
msgpack==1.1.2
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psycopg==3.3.3
psycopg-binary==3.3.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydyf==0.12.1
PyJWT==2.11.0
pyparsing==3.3.2
pyphen==0.17.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.4.1
